Make valid VIF types configurable
Port binding allow to specify the list of valid VIF types to be used
for port binding. This allow to restrict selection to a subset and to
give preference hints to ML2 drivers.

This list of strings has to be optionally read from driver
configuration file and from local.conf.

NOTE: this change was split from below change to make it
lighter for code review:

https://review.openstack.org/#/c/273390/

Change-Id: Ifce3ff93527976f7469e648cedb49d015c1ba08b
Closes-Bug: #1552172